# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Silk Way API",
        "description": "REST API для мобильного приложения Silk Way. Авторизация через Bearer токен (Sanctum).",
        "contact": {
            "email": "admin@silkway.kz"
        },
        "version": "1.9.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Получить профиль текущего пользователя",
                "operationId": "ba77fb1a2bb44ad4937210a8e82bc344",
                "responses": {
                    "200": {
                        "description": "Данные профиля",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Обновить профиль текущего пользователя",
                "operationId": "048d08d14976ae7214bfc54d17a659f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "email": {
                                        "description": "Только для warehouse_employee / admin",
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "password": {
                                        "description": "Только для warehouse_employee / admin; не заполнять — пароль не меняется",
                                        "type": "string",
                                        "format": "password",
                                        "example": "newSecret123"
                                    },
                                    "password_confirmation": {
                                        "description": "Только для warehouse_employee / admin",
                                        "type": "string",
                                        "format": "password",
                                        "example": "newSecret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Профиль обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Профиль обновлён."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "example": "ivan@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "warehouse_employee",
                            "driver"
                        ]
                    },
                    "is_approved": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введите токен (без слова Bearer — оно добавляется автоматически)",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}